Move whirl related methods to results
All methods related to whirl calculation are now associated with
ModalResults.
